+++ Study_Plan.txt
@@ -4,5 +4,5 @@ CS2101 CS2113T CG2271 GER1000 GET1013
 CG2023 ST2334 CG2027 CG2028 GEH1036
 CP3880 CS3230
 EG2401A
-
+CS1123
