+++ src/main/java/gazeeebo/commands/capCalculator/DeleteCAPCommand.java
@@ -1,7 +1,7 @@
 package gazeeebo.commands.capCalculator;
 
 import gazeeebo.UI.Ui;
-import gazeeebo.parsers.CAPCommandParser;
+import gazeeebo.parser.CAPCommandParser;
 
 import java.io.IOException;
 import java.util.ArrayList;
