+++ src/main/java/gazeeebo/parser/CAPCommandParser.java
@@ -1,5 +1,5 @@
 
-package gazeeebo.parsers;
+package gazeeebo.parser;
 
 import gazeeebo.TriviaManager.TriviaManager;
 import gazeeebo.UI.Ui;
