+++ src/main/java/gazeeebo/parser/PlacesCommandParser.java
@@ -1,6 +1,6 @@
 //@@author jessteoxizhi
 
-package gazeeebo.parsers;
+package gazeeebo.parser;
 
 import gazeeebo.commands.help.HelpCommand;
 import gazeeebo.commands.places.*;
@@ -71,19 +71,19 @@ public class PlacesCommandParser extends Command {
                 } else if (ui.fullCommand.equals("help")) {
                     new HelpCommand().execute(list,ui,storage,commandStack,deletedTask,triviaManager);
                 } else if (ui.fullCommand.equals("esc")) {
-                    System.out.println("Going back to Main Menu...\n" +
-                            "Content Page:\n" +
-                            "------------------ \n" +
-                            "1. help\n" +
-                            "2. contacts\n" +
-                            "3. expenses\n" +
-                            "4. places\n" +
-                            "5. tasks\n" +
-                            "6. cap\n" +
-                            "7. spec\n" +
-                            "8. moduleplanner\n" +
-                            "9. notes\n" +
-                            "To exit: bye\n");
+                    System.out.println("Going back to Main Menu...\n"
+                            + "Content Page:\n"
+                            + "------------------ \n"
+                            + "1. help\n"
+                            + "2. contacts\n"
+                            + "3. expenses\n"
+                            + "4. places\n"
+                            + "5. tasks\n"
+                            + "6. cap\n"
+                            + "7. spec\n"
+                            + "8. moduleplanner\n"
+                            + "9. notes\n"
+                            + "To exit: bye\n");
                     isExitFromPlaces = true;
                 } else if (ui.fullCommand.equals("undo")) {
                     places = UndoPlacesCommand.undoPlaces(places,oldplaces,storage);
